简化kw歌词获取 && kw歌词解密逻辑移动
将酷我歌词解密从主进程 native/IPC 方案迁移到渲染进程纯 JS 实现，
移除 kw_decodeLyric 主进程入口与 handle_kw_decode_lyric IPC 通道。
