+++ src/common/ipcNames.ts
@@ -80,7 +80,6 @@ const modules = {
 
     // lang_s2t: 'lang_s2t',
 
-    handle_kw_decode_lyric: 'handle_kw_decode_lyric',
     handle_tx_decode_lyric: 'handle_tx_decode_lyric',
     get_lyric_info: 'get_lyric_info',
     set_lyric_info: 'set_lyric_info',
