+++ src/main/modules/winMain/rendererEvent/index.ts
@@ -3,7 +3,6 @@ import { registerRendererEvents as list } from '@main/modules/commonRenderers/li
 import { registerRendererEvents as dislike } from '@main/modules/commonRenderers/dislike'
 import app, { sendConfigChange } from './app'
 import hotKey from './hotKey'
-import kw_decodeLyric from './kw_decodeLyric'
 import tx_decodeLyric from './tx_decodeLyric'
 import userApi from './userApi'
 import sync from './sync'
@@ -30,7 +29,6 @@ export default () => {
   dislike(sendEvent)
   app()
   hotKey()
-  kw_decodeLyric()
   tx_decodeLyric()
   userApi()
   sync()
