+++ src/renderer/utils/musicSdk/kw/util.js
@@ -1,95 +1,55 @@
-// import { httpGet, httpFetch } from '../../request'
-import { WIN_MAIN_RENDERER_EVENT_NAME } from '@common/ipcNames'
-import { rendererInvoke } from '@common/rendererIpc'
 import { createCipheriv, createDecipheriv } from 'crypto'
 import { toMD5 } from '../utils'
+import { inflate } from 'zlib'
+import iconv from 'iconv-lite'
 
-// const kw_token = {
-//   token: null,
-//   isGetingToken: false,
-// }
 
-// const translationMap = {
-//   "{'": '{"',
-//   "'}\n": '"}',
-//   "'}": '"}',
-//   "':'": '":"',
-//   "','": '","',
-//   "':{'": '":{"',
-//   "':['": '":["',
-//   "'}],'": '"}],"',
-//   "':[{'": '":[{"',
-//   "'},'": '"},"',
-//   "'},{'": '"},{"',
-//   "':[],'": '":[],"',
-//   "':{},'": '":{},"',
-//   "'}]}": '"}]}',
-// }
+const handleInflate = async(data) => {
+  return new Promise((resolve, reject) => {
+    inflate(data, (err, result) => {
+      if (err) {
+        reject(err)
+        return
+      }
+      resolve(result)
+    })
+  })
+}
 
-// export const objStr2JSON = str => {
-//   return JSON.parse(str.replace(/(^{'|'}\n$|'}$|':'|','|':\[{'|'}\],'|':{'|'},'|'},{'|':\['|':\[\],'|':{},'|'}]})/g, s => translationMap[s]))
-// }
+const buf_key = Buffer.from('yeelion')
+const buf_key_len = buf_key.length
+
 
 export const objStr2JSON = str => {
   return JSON.parse(str.replace(/('(?=(,\s*')))|('(?=:))|((?<=([:,]\s*))')|((?<={)')|('(?=}))/g, '"'))
 }
-
-
 export const formatSinger = rawData => rawData.replace(/&/g, '、')
-
-export const matchToken = headers => {
-  try {
-    return headers['set-cookie'][0].match(/kw_token=(\w+)/)[1]
-  } catch (err) {
-    return null
+export const decodeLyric = async(rawData, isGetLyricx) => {
+  const buf = Buffer.isBuffer(rawData) ? rawData : Buffer.from(rawData)
+  if (buf.toString('utf8', 0, 10).toLowerCase() !== 'tp=content') return ''
+  const lrcData = await handleInflate(buf.subarray(buf.indexOf('\r\n\r\n') + 4))
+  if (!isGetLyricx) return iconv.decode(lrcData, 'utf8')
+  const buf_str = Buffer.from(lrcData.toString(), 'base64')
+  const buf_str_len = buf_str.length
+  const output = new Uint8Array(buf_str_len)
+  let i = 0
+  while (i < buf_str_len) {
+    let j = 0
+    while (j < buf_key_len && i < buf_str_len) {
+      output[i] = buf_str[i] ^ buf_key[j]
+      i++
+      j++
+    }
   }
+
+  return iconv.decode(Buffer.from(output), 'utf8')
 }
 
-// const wait = time => new Promise(resolve => setTimeout(() => resolve(), time))
-
-
-// export const getToken = (retryNum = 0) => new Promise((resolve, reject) => {
-//   if (retryNum > 2) return Promise.reject(new Error('try max num'))
-
-//   if (kw_token.isGetingToken) return wait(1000).then(() => getToken(retryNum).then(token => resolve(token)))
-//   if (kw_token.token) return resolve(kw_token.token)
-//   kw_token.isGetingToken = true
-//   httpGet('http://www.kuwo.cn/', (err, resp) => {
-//     kw_token.isGetingToken = false
-//     if (err) return getToken(++retryNum)
-//     if (resp.statusCode != 200) return reject(new Error('获取失败'))
-//     const token = kw_token.token = matchToken(resp.headers)
-//     resolve(token)
-//   })
-// })
-
-export const decodeLyric = base64Data => rendererInvoke(WIN_MAIN_RENDERER_EVENT_NAME.handle_kw_decode_lyric, base64Data)
-
-// export const tokenRequest = async(url, options = {}) => {
-//   let token = kw_token.token
-//   if (!token) token = await getToken()
-//   if (!options.headers) {
-//     options.headers = {
-//       Referer: 'http://www.kuwo.cn/',
-//       csrf: token,
-//       cookie: 'kw_token=' + token,
-//     }
-//   }
-//   const requestObj = httpFetch(url, options)
-//   requestObj.promise = requestObj.promise.then(resp => {
-//     // console.log(resp)
-//     if (resp.statusCode == 200) {
-//       kw_token.token = matchToken(resp.headers)
-//     }
-//     return resp
-//   })
-//   return requestObj
-// }
 
 export const lrcTools = {
   rxps: {
-    wordLine: /^(\[\d{1,2}:.*\d{1,4}\])\s*(\S+(?:\s+\S+)*)?\s*/,
-    tagLine: /\[(ver|ti|ar|al|offset|by|kuwo):\s*(\S+(?:\s+\S+)*)\s*\]/,
+    wordLine: /^(\[\d{1,2}:.*\d{1,4}])\s*(\S+(?:\s+\S+)*)?\s*/,
+    tagLine: /\[(ver|ti|ar|al|offset|by|kuwo):\s*(\S+(?:\s+\S+)*)\s*]/,
     wordTimeAll: /<(-?\d+),(-?\d+)(?:,-?\d+)?>/g,
     wordTime: /<(-?\d+),(-?\d+)(?:,-?\d+)?>/,
   },
@@ -111,7 +71,6 @@ export const lrcTools = {
         }
 
         prevWord.newTimeStr = `<${prevWord.startTime},${prevWord.endTime - prevWord.startTime}>`
-        // console.log(prevWord)
       }
     }
     return {
@@ -131,7 +90,6 @@ export const lrcTools = {
       }
       const wordTimes = words.match(this.rxps.wordTimeAll)
       if (!wordTimes) return
-      // console.log(wordTimes)
       let preTimeInfo
       for (const timeStr of wordTimes) {
         const result = this.rxps.wordTime.exec(timeStr)
@@ -145,7 +103,7 @@ export const lrcTools = {
     }
     result = this.rxps.tagLine.exec(line)
     if (!result) return
-    if (result[1] == 'kuwo') {
+    if (result[1] === 'kuwo') {
       let content = result[2]
       if (content != null && content.includes('][')) {
         content = content.substring(0, content.indexOf(']['))
@@ -153,7 +111,7 @@ export const lrcTools = {
       const valueOf = parseInt(content, 8)
       this.offset = Math.trunc(valueOf / 10)
       this.offset2 = Math.trunc(valueOf % 10)
-      if (this.offset == 0 || Number.isNaN(this.offset) || this.offset2 == 0 || Number.isNaN(this.offset2)) {
+      if (this.offset === 0 || Number.isNaN(this.offset) || this.offset2 === 0 || Number.isNaN(this.offset2)) {
         this.isOK = false
       }
     } else {
@@ -161,7 +119,6 @@ export const lrcTools = {
     }
   },
   parse(lrc) {
-    // console.log(lrc)
     const lines = lrc.split(/\r\n|\r|\n/)
     const tools = Object.create(this)
     tools.isOK = true
@@ -177,7 +134,6 @@ export const lrcTools = {
     if (!tools.lines.length) return ''
     let lrcs = tools.lines.join('\n')
     if (tools.tags.length) lrcs = `${tools.tags.join('\n')}\n${lrcs}`
-    // console.log(lrcs)
     return lrcs
   },
 }
